+++ src/components/Unit/Presentation/ParentCard.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import {Col, Row} from 'rivet-react'
-import { IEntity } from "../types";
-import { ParentUnitIcon } from '../icons';
+import { IEntity } from "../../types";
+import { ParentUnitIcon } from '../../icons';
 
 interface IProps {
     parent?: IEntity
